Move config to pygmt/src/config.py
