+++ pygmt/__init__.py
@@ -25,11 +25,12 @@ from pkg_resources import get_distribution
 # Import modules to make the high-level GMT Python API
 from pygmt import datasets
 from pygmt.figure import Figure
-from pygmt.modules import GMTDataArrayAccessor, config
+from pygmt.modules import GMTDataArrayAccessor
 from pygmt.session_management import begin as _begin
 from pygmt.session_management import end as _end
 from pygmt.src import (
     blockmedian,
+    config,
     grd2cpt,
     grdcut,
     grdfilter,
