+++ pygmt/src/__init__.py
@@ -6,6 +6,7 @@ from pygmt.src.basemap import basemap
 from pygmt.src.blockmedian import blockmedian
 from pygmt.src.coast import coast
 from pygmt.src.colorbar import colorbar
+from pygmt.src.config import config
 from pygmt.src.contour import contour
 from pygmt.src.grd2cpt import grd2cpt
 from pygmt.src.grdcontour import grdcontour
